fix Serialize Exception for cached object
