+++ discovery-server/src/main/java/app/metatron/discovery/common/oauth/token/cache/WhitelistTokenCacheRepository.java
@@ -23,8 +23,6 @@ import org.springframework.cache.annotation.CachePut;
 import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Repository;
 
-import java.io.Serializable;
-
 /**
  *
  */
@@ -56,59 +54,13 @@ public class WhitelistTokenCacheRepository {
     cacheManager.getCache("token-whitelist-cache").evict(username + "|" + clientId);
   }
 
-  public void removeWhitelistTokenByCachedAccessToken(AccessTokenCacheRepository.CachedAccessToken cachedAccessToken){
+  public void removeWhitelistTokenByCachedAccessToken(CachedAccessToken cachedAccessToken){
     if(cachedAccessToken != null){
-      WhitelistTokenCacheRepository.CachedWhitelistToken cachedWhitelistToken
+      CachedWhitelistToken cachedWhitelistToken
           = this.getCachedWhitelistToken(cachedAccessToken.getUsername(), cachedAccessToken.getClientId());
       if(cachedWhitelistToken != null && cachedAccessToken.getToken().equals(cachedWhitelistToken.getToken())){
         this.removeWhitelistToken(cachedAccessToken.getUsername(), cachedAccessToken.getClientId());
       }
-    }
-  }
-
-  public class CachedWhitelistToken implements Serializable {
-    public String token;
-    public String username;
-    public String clientId;
-    public String userHost;
-
-    public CachedWhitelistToken(String token, String username, String clientId, String userHost) {
-      this.token = token;
-      this.username = username;
-      this.clientId = clientId;
-      this.userHost = userHost;
-    }
-
-    public String getToken() {
-      return token;
-    }
-
-    public void setToken(String token) {
-      this.token = token;
-    }
-
-    public String getUsername() {
-      return username;
-    }
-
-    public void setUsername(String username) {
-      this.username = username;
-    }
-
-    public String getClientId() {
-      return clientId;
-    }
-
-    public void setClientId(String clientId) {
-      this.clientId = clientId;
-    }
-
-    public String getUserHost() {
-      return userHost;
-    }
-
-    public void setUserHost(String userHost) {
-      this.userHost = userHost;
     }
   }
 }
